requirements.txt
discord.py
python-dotenv
notion-client
google-generativeai
aiohttp
